Move api to exported library folder
